refactor(security): extract TrustedHostPolicy to break router-mcp cycle
