+++ autumn/src/router.rs
@@ -644,7 +644,7 @@ fn build_router_pre_state(
             cors: config.cors.clone(),
             // The same-origin shortcut is gated on the app's trusted-Host
             // policy so it can't be abused for DNS rebinding.
-            trusted_hosts: TrustedHostPolicy::from_config(config),
+            trusted_hosts: crate::security::TrustedHostPolicy::from_config(config),
             tenant_header,
             // Forward the configured CSRF header (default `x-csrf-token`) so
             // customized CsrfConfig::token_header deployments work via MCP.
@@ -2275,7 +2275,7 @@ fn apply_middleware(
         };
 
     router = apply_cors_middleware(router, config);
-    let trusted_host_policy = TrustedHostPolicy::from_config(config);
+    let trusted_host_policy = crate::security::TrustedHostPolicy::from_config(config);
     router = router.layer(axum::middleware::from_fn(move |req, next| {
         trusted_host_middleware(req, next, trusted_host_policy.clone())
     }));
@@ -2539,7 +2539,7 @@ fn apply_layers_in_registration_order(
 async fn trusted_host_middleware(
     req: Request<axum::body::Body>,
     next: Next,
-    policy: TrustedHostPolicy,
+    policy: crate::security::TrustedHostPolicy,
 ) -> axum::response::Response {
     let path = req.uri().path();
     if (req.method() == http::Method::GET || req.method() == http::Method::HEAD)
@@ -2553,7 +2553,7 @@ async fn trusted_host_middleware(
         .get(http::header::HOST)
         .and_then(|v| v.to_str().ok());
     let raw_host = authority.or(host_header);
-    let parsed_host = raw_host.and_then(extract_host_without_port);
+    let parsed_host = raw_host.and_then(crate::security::extract_host_without_port);
     let host = parsed_host
         .map(str::to_ascii_lowercase)
         .map(|h| h.trim_end_matches('.').to_owned())
@@ -2581,47 +2581,6 @@ async fn trusted_host_middleware(
             body,
         )
             .into_response()
-    }
-}
-
-pub fn extract_host_without_port(header: &str) -> Option<&str> {
-    let host = header.trim();
-    if host.is_empty() {
-        return None;
-    }
-    if host.starts_with('[') {
-        let end = host.find(']')?;
-        let literal = host.get(1..end)?;
-        if literal.is_empty() || literal.parse::<std::net::IpAddr>().is_err() {
-            return None;
-        }
-
-        let remainder = host.get(end + 1..)?;
-        if remainder.is_empty() {
-            return Some(literal);
-        }
-
-        let maybe_port = remainder.strip_prefix(':')?;
-        if !maybe_port.is_empty() && maybe_port.chars().all(|c| c.is_ascii_digit()) {
-            return Some(literal);
-        }
-
-        return None;
-    }
-    let Some((candidate, maybe_port)) = host.rsplit_once(':') else {
-        return Some(host);
-    };
-    if candidate.contains(':') {
-        // unbracketed IPv6 literal; keep host verbatim
-        return Some(host);
-    }
-    if !maybe_port.is_empty()
-        && maybe_port.chars().all(|c| c.is_ascii_digit())
-        && !candidate.is_empty()
-    {
-        Some(candidate)
-    } else {
-        None
     }
 }
 
@@ -3756,7 +3715,7 @@ mod tests {
 
         let wiring = crate::mcp::McpWiring {
             cors: crate::config::CorsConfig::default(),
-            trusted_hosts: TrustedHostPolicy::from_config(&config),
+            trusted_hosts: crate::security::TrustedHostPolicy::from_config(&config),
             tenant_header: None,
             csrf_header: "x-csrf-token".to_owned(),
             envelope_rate_limited: false,
@@ -6113,77 +6072,6 @@ mod trusted_host_tests {
         assert!(super::custom_layers_require_fail_closed_idempotency(&gate));
         // An empty set requires no fail-closed behaviour.
         assert!(!super::custom_layers_require_fail_closed_idempotency(&[]));
-    }
-}
-#[derive(Clone, Debug)]
-pub struct TrustedHostPolicy {
-    rules: Arc<Vec<String>>,
-    allow_any: bool,
-    allow_missing_host: bool,
-    probe_bypass_paths: Arc<std::collections::HashSet<String>>,
-}
-
-impl TrustedHostPolicy {
-    pub fn from_config(config: &AutumnConfig) -> Self {
-        let mut rules: Vec<String> = config
-            .security
-            .trusted_hosts
-            .hosts
-            .iter()
-            .map(|h| h.trim().to_ascii_lowercase())
-            .map(|h| h.trim_end_matches('.').to_owned())
-            .filter(|h| !h.is_empty())
-            .collect();
-        let is_production = matches!(config.profile.as_deref(), Some("prod" | "production"));
-        if !is_production {
-            rules.extend(
-                ["localhost", "127.0.0.1", "::1"]
-                    .into_iter()
-                    .map(std::borrow::ToOwned::to_owned),
-            );
-        }
-        let allow_any = rules.iter().any(|h| h == "*");
-        let probe_bypass_paths = std::collections::HashSet::from([
-            config.health.path.clone(),
-            config.health.live_path.clone(),
-            config.health.ready_path.clone(),
-            config.health.startup_path.clone(),
-            crate::actuator::actuator_route_path(&config.actuator.prefix, "/health"),
-        ]);
-        Self {
-            rules: Arc::new(rules),
-            allow_any,
-            allow_missing_host: !is_production,
-            probe_bypass_paths: Arc::new(probe_bypass_paths),
-        }
-    }
-
-    /// Whether a request carrying no usable `Host` is allowed through. Mirrors
-    /// `trusted_host_middleware`'s missing-host branch for callers (e.g. the MCP
-    /// envelope) that enforce the policy outside that middleware.
-    ///
-    /// Only the `mcp` feature consumes this today; gated so default-feature
-    /// builds don't flag it as dead code.
-    #[cfg(feature = "mcp")]
-    pub const fn allows_missing_host(&self) -> bool {
-        self.allow_missing_host
-    }
-
-    pub fn allows_host(&self, host: &str) -> bool {
-        if self.allow_any {
-            return true;
-        }
-        self.rules.iter().any(|rule| {
-            rule.strip_prefix('.').map_or_else(
-                || host == rule,
-                |suffix| {
-                    host == suffix
-                        || host
-                            .strip_suffix(suffix)
-                            .is_some_and(|prefix| prefix.ends_with('.'))
-                },
-            )
-        })
     }
 }
