+++ autumn/src/security/mod.rs
@@ -143,6 +143,7 @@ pub(crate) mod csrf;
 pub(crate) mod headers;
 pub mod proxy;
 pub mod rate_limit;
+pub mod trusted_host;
 pub(crate) mod trusted_proxies;
 
 // Re-export commonly used types at the module level.
@@ -163,4 +164,5 @@ pub use csrf::{CsrfFormField, CsrfLayer, CsrfToken, CsrfTokenHeader};
 pub use headers::{CspNonce, SecurityHeadersLayer};
 pub use proxy::TrustedProxy;
 pub use rate_limit::{RateLimitExempt, RateLimitLayer, RateLimitOverride, RateLimitPrincipal};
+pub use trusted_host::{TrustedHostPolicy, extract_host_without_port};
 pub use trusted_proxies::{ProxyResolver, ResolvedClientIdentity, TrustedProxiesLayer};
