+++ plan.md
@@ -1,0 +1,1 @@
+1. Break `mcp` -> `router` circular dependency by moving `TrustedHostPolicy` and `extract_host_without_port` out of `router.rs` and into a smaller module. Looking closely at `router.rs`, `TrustedHostPolicy` is related to `security` or `config` or maybe a new module like `host_policy.rs` but the easiest way might be moving `TrustedHostPolicy` and `extract_host_without_port` to `security/trusted_host.rs` or `security::trusted_host`.
